# EICR_App/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CertMate (EICR-oMatic 3000) API
  version: 1.0.0
  description: |
    Backend API for CertMate — automated EICR/EIC electrical certificate creation
    for electrical inspectors. Powered by Express/Node.js with ES modules.

    ## Authentication
    Most endpoints require a Bearer JWT token obtained via `POST /api/auth/login`.
    Pass the token in the `Authorization` header: `Bearer <token>`.

    ## Certificate Types
    - **EICR** — Electrical Installation Condition Report (periodic inspection)
    - **EIC** — Electrical Installation Certificate (new installations)
  contact:
    name: CertMate Support
  license:
    name: Proprietary
servers:
  - url: https://certomatic3000.co.uk
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /api/job/{userId}/{jobId}/generate-pdf:
    post:
      tags: [PDF]
      summary: Generate PDF certificate for a job
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF generated and returned
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Job not found
...
  /api/ocr/certificate:
    post:
      tags: [OCR]
      summary: Extract data from an existing certificate via OCR
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF, JPG, JPEG, or PNG file
      responses:
        "200":
          description: Extraction successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    description: Extracted certificate data in CertMate format
                  meta:
                    type: object
                    properties:
                      model:
                        type: string
                      tokens:
                        type: integer
                      source_file:
                        type: string
        "400":
          description: No file or unsupported type
  /api/ocr/create-job:
    post:
      tags: [OCR]
      summary: Create a job from OCR-extracted data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: object
                  description: Extracted certificate data
                certificateType:
                  type: string
                  enum: [EICR, EIC]
      responses:
        "200":
          description: Job created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  jobId:
                    type: string
                  address:
                    type: string

  # ============= Settings =============
...